pyproject.toml
[project]
name = "downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ffmpeg-python==0.2.0",
    "opencv-python>=4.11.0.86",
    "prompt-toolkit>=3.0.51",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "python-ffmpeg>=2.0.12",
    "selenium>=4.32.0",
]
